requirements.txt
uvicorn
fastapi
transformers
python-dotenv
google-generativeai
torch
